+++ src/Savvyio.EventDriven/IIntegrationEvent.cs
@@ -1,7 +1,7 @@
 ﻿namespace Savvyio.EventDriven
 {
     /// <summary>
-    /// A marker interface that specifies something that happened when an Aggregate was successfully persisted and you want other subsystems (out-process/inter-application) to be made aware of.
+    /// A marker interface that specifies something that happened when an object was successfully persisted and you want other subsystems (out-process/inter-application) to be made aware of.
     /// </summary>
     public interface IIntegrationEvent : IRequest, IMetadata
     {