+++ src/Savvyio.EventDriven/IntegrationEvent.cs
@@ -13,7 +13,6 @@ namespace Savvyio.EventDriven
         /// Initializes a new instance of the <see cref="IntegrationEvent"/> class.
         /// </summary>
         /// <param name="eventId">The optional identifier of the event. Default is an auto-generated UUID.</param>
-        /// <param name="type">The optional type of the event. Default is the type of this instance.</param>
         /// <param name="metadata">The optional metadata to merge with this instance.</param>
         /// <remarks>
         /// The following table shows the initial metadata values for an instance of <see cref="IntegrationEvent"/>.
@@ -30,17 +29,12 @@ namespace Savvyio.EventDriven
         ///         <term><see cref="MetadataDictionary.Timestamp"/></term>
         ///         <description><c>DateTime.UtcNow</c></description>
         ///     </item>
-        ///     <item>
-        ///         <term><see cref="MetadataDictionary.MemberType"/></term>
-        ///         <description><c>type ?? GetType()</c></description>
-        ///     </item>
         /// </list>
         /// </remarks>
-        protected IntegrationEvent(string eventId = null, Type type = null, IMetadata metadata = null)
+        protected IntegrationEvent(string eventId = null, IMetadata metadata = null)
         {
             this.SetEventId(eventId ?? Guid.NewGuid().ToString("N"));
             this.SetTimestamp();
-            this.SetMemberType(type ?? GetType());
             this.MergeMetadata(metadata);
         }
     }