+++ src/Savvyio.Extensions.SimpleQueueService/Commands/AmazonCommandQueueOptions.cs
@@ -1,0 +1,10 @@
+﻿namespace Savvyio.Extensions.SimpleQueueService.Commands
+{
+    /// <summary>
+    /// Configuration options for <see cref="AmazonCommandQueue"/>.
+    /// </summary>
+    /// <seealso cref="AmazonMessageOptions"/>
+    public class AmazonCommandQueueOptions : AmazonMessageOptions
+    {
+    }
+}