+++ test/Savvyio.Core.Tests/Assets/RootDummyRequest.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Savvyio.Assets
+{
+    public record RootDummyRequest(Guid Uuid, int Number, bool Hidden = false) : DummyRequest(Uuid, Number)
+    {
+    }
+}